This is a test to see how easy it is to copy and paste a Gist Tutorial I wrote into the new jekyll Blog format
I'm specifically testing to see if the images transfer.  That's always been the most annoying thing about documentation for me: managing the images. 

I have a fairly simple workflow to make tutorials full of images like this using gists.  The most annoying part of tutorial creation is making sure that appropirate images are added, and if  copying the tutorial from the gist to this format is a simple matter then it will be worthwhile to write up the process of how to create these gist tutorials.   (short version: you can't drag an image into a gist, but you can drag an image into a comment.  Github will upload that image into the comment and give you a markdown link.  Then you can copy the link into your gist tutorial and have nice images.)