Update docs to be in sync with current public release of Talon
As of the Talon 0.1.0 release, the public release is now the new api (formerly
only available with the beta release). This change updates the docs to reflect
that change, which is mostly removing references to the beta where it's no
longer needed. There is still a beta version of Talon, but the only current
difference is access to extra speech engines and slightly newer code.

The changelog for Talon is here: https://talonvoice.com/dl/latest/changelog.html.