Remove "Many commands work badly" section
This section as written is problematic. It unnecessarily negatively phrased and
may demotivate users. The phrasing "Many commands work badly" may suggest a
statement of fact instead of a fixable situation the user might be in. Remove
it for now - we can add useful bits back later in a different way.