removed advanced_talon.md
This is another grab-bag page that has a lot of outdated information.

First, the most recommendended and heavily maintained repository is knausj.  All other repositories are individually maintained, and optimized for particular operating systems.  I think people should find them via the slack, where they can be guided through a setup process, and not through this wiki.

Mousegrid is specific to knausj, the instructions don't say how to turn it on, and the instructions for activating it are in the readme under the settings section.

notification for what Talon understands seems like a strange thing to give access to non-programmers.

section 4, with all of it's 'this will be broken tomorrow!' disclamers, is also not the sort of evergreen content that belongs in a wiki.