Change title of third-party software page and add to sidebar
The previous title implied that the page was about integrating software with
Talon, which sounds more like writing Talon configuration. The title was a bit
misleading, so change it to be about software that pairs nicely with Talon.
Also, add this page to the sidebar and fix a couple nits with language and
punctuation.