Move the speech engine table to its own page
This change simplifies the getting started page to prefer the Conformer speech
engine, but still links to additional information about the speech engines. The
intent is to reduce friction for new users who may not need all the speech
engine detail. This change also adds the speech engines page to the nav.