Update talon_user_file_sets.md to include Talon AI Tools
A fair amount of people have begun to use my AI tool repository for fixing dictation errors or using Copilot alongside cursorless.  That latter feature was contributed to my repository by pokey. As a result I think it is a good idea to include it here on the wiki.  I put it second in the list because it is second in the list in terms of popularity according to GitHub stars. 

 The repository link can be found here: https://github.com/C-Loftus/talon-ai-tools