add an alternative deploy to github pages (#345)
add an alternative deployment to github pages

This change coexists with the current deploy to Netlify. Once merged, any PR merged to main will automatically deploy to both Netlify and GitHub pages. At that point, if we are happy with the GitHub pages version, talon.wiki domain can be updated.